backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
openai
marker-pdf
